trial 1

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ long int li;
 long long int lli;
 float f;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(c));
-printf("Size of an int: %lu byte(s)\n", (unsigned long) sizeof(i));
-printf("size of a long int: %lu byte(s)\n", (unsigned long) sizeof(li));
-printf("size of a long long int: %lu byte(s)\n", (unsigned long) sizeof(lli));
-printf("size of a float: %lu byte(s)\n", (unsigned long) sizeof(f));
+printf("Size of a char: %d byte(s)\n", sizeof(c));
+printf("Size of an int: %d byte(s)\n", sizeof(i));
+printf("size of a long int: %d byte(s)\n", sizeof(li));
+printf("size of a long long int: %d byte(s)\n", sizeof(lli));
+printf("size of a float: %d byte(s)\n", sizeof(f));
 return (0);
 }
